Removes unnecessary files from workers/test_volume resulting in >6GB no longer sitting around on disk all the time.

+++ workers/run_tests.sh
@@ -66,11 +66,12 @@ test_data_repo="https://s3.amazonaws.com/data-refinery-test-assets"
 
 if [[ -z $tag || $tag == "salmon" ]]; then
     # Download "salmon quant" test data
-    echo "Downloading 'salmon quant' test data..."
-    rm -rf $volume_directory/salmon_tests/
-
-    wget -q -O $volume_directory/salmon_tests.tar.gz $test_data_repo/salmon_tests.tar.gz
-    tar xzf $volume_directory/salmon_tests.tar.gz -C $volume_directory
+    if [ ! -e $volume_directory/salmon_tests ]; then
+        echo "Downloading 'salmon quant' test data..."
+        wget -q -O $volume_directory/salmon_tests.tar.gz $test_data_repo/salmon_tests.tar.gz
+        tar xzf $volume_directory/salmon_tests.tar.gz -C $volume_directory
+        rm $volume_directory/salmon_tests.tar.gz
+    fi
 
     # Download salmontools test data
     rm -rf $volume_directory/salmontools/
